requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
dominate==2.9.1
flask==3.1.3
flask-bootstrap==3.3.7.1
flask-login==0.6.3
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==3.0.3
pip==25.0.1
python-dotenv==1.2.1
sqlalchemy==2.0.47
typing-extensions==4.15.0
visitor==0.1.3
waitress==3.0.2
werkzeug==3.1.6
